deltaTime 가지고 스피드 조절

+++ prac04/player.h
@@ -8,5 +8,5 @@ public:
 	float playerX;
 	float playerY;
 
-	float speed;
+	int speed;
 };
